+++ doc/whats-new.rst
@@ -29,9 +29,10 @@ Enhancements
 Bug fixes
 ~~~~~~~~~
 
+- Dataset.copy(deep=True) now creates a deep copy of the attrs (:issue:`2835`).
+  By `Andras Gefferth <https://github.com/kefirbandi>`_.
 - ``swap_dims`` would create incorrect ``indexes`` (:issue:`2842`).
   By `Stephan Hoyer <https://github.com/shoyer>`_.
-
 
 .. _whats-new.0.12.0: